- Fixed the crashing issue in the font writer : type mismatch (was sending an array of integers when the function was expecting an array of shorts).

+++ src/font_desc.h
@@ -3,6 +3,6 @@
 /* Font descriptor headers */
 
 extern const char tiny_font_glyph[44];
-extern const int tiny_font_x_pos[44];
+extern const short tiny_font_x_pos[44];
 
 #endif
